test: split Forgejo project service coverage
Generated-By: looper 0.10.2 (runner=fixer, agent=codex)
